+++ src/main/java/conversionRequest/CardMdCycAcctHdevn.java
@@ -14,7 +14,7 @@ public class CardMdCycAcctHdevn {
     private String TRANS_TYPE_CODE;
     private String TRANS_TYPE;
     private Integer TRANS_CNT;
-    private Double TRANS_AMT;
+    private Integer TRANS_AMT;
     private Date MYETL_DATE;
 
     public String getXACCOUNT() {
@@ -73,11 +73,11 @@ public class CardMdCycAcctHdevn {
         this.TRANS_CNT = TRANS_CNT;
     }
 
-    public Double getTRANS_AMT() {
+    public Integer getTRANS_AMT() {
         return TRANS_AMT;
     }
 
-    public void setTRANS_AMT(Double TRANS_AMT) {
+    public void setTRANS_AMT(Integer TRANS_AMT) {
         this.TRANS_AMT = TRANS_AMT;
     }
